Fix support for osx
